Capstone
the application aims to automatically adjust the status of a mobile
phone based on the user’s location, the time of day, and activity of
the user. It has been observed that people want to have more control
over when they receive cell phone calls, depending on their current
context. For example, they prefer not to take unexpected calls during a
meeting with their boss. This project identifies three different
environments: at work, at home and on the go. When at work, the phone
is set to be in silent or vibration mode for meetings based on the
calendar. The application also gives suggestions for restaurants near
the office during lunchtime. When at home, the application will list
movies playing for the next few hours in cinemas nearby. When on the
go, the application will give directions from the phone’s current
location to points of interest, the distance to the desired
destination, and estimated time of travel.